move some files for better organization. Cart projection value setted on the first rendering
